backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
PyMuPDF==1.23.8
aiofiles==23.2.1
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23

# Testing and code quality
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.0
flake8==6.1.0
